Reapply #2601
