--- meilisearch-lib/Cargo.toml
@@ -1,85 +1,0 @@
-[package]
-name = "meilisearch-lib"
-version = "0.29.1"
-edition = "2021"
-
-[dependencies]
-actix-web = { version = "4.2.1", default-features = false }
-anyhow = { version = "1.0.65", features = ["backtrace"] }
-async-stream = "0.3.3"
-async-trait = "0.1.57"
-atomic_refcell = "0.1.8"
-byte-unit = { version = "4.0.14", default-features = false, features = ["std", "serde"] }
-bytes = "1.2.1"
-clap = { version = "4.0.9", features = ["derive", "env"] }
-crossbeam-channel = "0.5.6"
-csv = "1.1.6"
-derivative = "2.2.0"
-either = { version = "1.8.0", features = ["serde"] }
-flate2 = "1.0.24"
-fs_extra = "1.2.0"
-fst = "0.4.7"
-futures = "0.3.24"
-futures-util = "0.3.24"
-http = "0.2.8"
-indexmap = { version = "1.9.1", features = ["serde-1"] }
-itertools = "0.10.5"
-lazy_static = "1.4.0"
-log = "0.4.17"
-meilisearch-auth = { path = "../meilisearch-auth" }
-meilisearch-types = { path = "../meilisearch-types" }
-milli = { git = "https://github.com/meilisearch/milli.git", tag = "v0.34.0", default-features = false }
-mime = "0.3.16"
-num_cpus = "1.13.1"
-obkv = "0.2.0"
-once_cell = "1.15.0"
-page_size = "0.4.2"
-parking_lot = "0.12.1"
-permissive-json-pointer = { path = "../permissive-json-pointer" }
-rand = "0.8.5"
-rayon = "1.5.3"
-regex = "1.6.0"
-reqwest = { version = "0.11.12", features = ["json", "rustls-tls"], default-features = false, optional = true }
-roaring = "0.10.1"
-rustls = "0.20.6"
-serde = { version = "1.0.145", features = ["derive"] }
-serde_json = { version = "1.0.85", features = ["preserve_order"] }
-siphasher = "0.3.10"
-slice-group-by = "0.3.0"
-sysinfo = "0.26.4"
-tar = "0.4.38"
-tempfile = "3.3.0"
-thiserror = "1.0.37"
-time = { version = "0.3.15", features = ["serde-well-known", "formatting", "parsing", "macros"] }
-tokio = { version = "1.21.2", features = ["full"] }
-uuid = { version = "1.1.2", features = ["serde", "v4"] }
-walkdir = "2.3.2"
-whoami = { version = "1.2.3", optional = true }
-index-scheduler = { path = "../index-scheduler" }
-index = { path = "../index" }
-file-store = { path = "../file-store" }
-
-[dev-dependencies]
-actix-rt = "2.7.0"
-meilisearch-types = { path = "../meilisearch-types", features = ["test-traits"] }
-mockall = "0.11.2"
-nelson = { git = "https://github.com/meilisearch/nelson.git", rev = "675f13885548fb415ead8fbb447e9e6d9314000a"}
-paste = "1.0.9"
-proptest = "1.0.0"
-proptest-derive = "0.3.0"
-
-[features]
-# all specialized tokenizations
-default = ["milli/default"]
-
-# chinese specialized tokenization
-chinese = ["milli/chinese"]
-
-# hebrew specialized tokenization
-hebrew = ["milli/hebrew"]
-
-# japanese specialized tokenization
-japanese = ["milli/japanese"]
-
-# thai specialized tokenization
-thai = ["milli/thai"]
